Add 'error messages' to process data
Python will now print what goes wrong in is_relevant. Also make the
capital D in the 'Direction'-import lowercase. I also increased the
distance threshold to 1000km so we don't have to constantly calculate
between latitude and km while testing. We must agree on a threshold and
set it later.